requirements.txt
torch
torchvision
timm
yacs
opencv-python

gitpython>=3.1.30
matplotlib>=3.2.2
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
tqdm>=4.64.0
tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0
setuptools>=65.5.1 # Snyk vulnerability fix
tritonclient[all]~=2.24.0
ultralytics  # HUB https://hub.ultralytics.com
moviepy
numpy
pycocotools
scikit-image
fastapi
uvicorn[standard]
SQLAlchemy
passlib
python-jose
bcrypt
python-multipart
email-validator
filterpy
protobuf
ipython

# live
imutils